Melhorias na componentização
