+++ semana9/projeto/astromatch/src/components/Header.js
@@ -1,21 +1,8 @@
 import React from "react";
-import styled from "styled-components";
 import FavoriteTwoToneIcon from "@material-ui/icons/FavoriteTwoTone";
 import GroupAddTwoToneIcon from "@material-ui/icons/GroupAddTwoTone";
+import { HeaderContainer, Logo } from "../components/styled";
 
-const HeaderContainer = styled.div`
-  height: 50px;
-  border-bottom: 1px solid grey;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  position: relative;
-  padding: 0px 10px;
-`;
-
-const Logo = styled.img`
-  width: 35%;
-`;
 
 const Header = (props) => {
